워프 기능 및 카메라 줌 이벤트 추가
- Asteroid.cpp: 소행성 이동 속도를 0으로 변경
- CameraSystem.cpp: 카메라 줌 이벤트 등록 추가
- Common.h: WARP_COMPONENT 추가
- Component.h: WarpComponent 구조체 정의
- Event.h: WarpStopEvent 및 CameraZoomEvent 추가
- Game.cpp: 게임 루프에서 RunProcess 호출 위치 수정
- InputSystem.cpp: 스페이스 키에 따른 속도 및 이벤트 처리 수정
- MoveSystem.cpp: 이동 로직 개선 및 slowVelocity 관리 추가
- Star.cpp: 별 초기화 시 WarpComponent 추가
- StarEffectSystem.cpp: 밝기 계산을 정수형으로 변환
- SystemManager.cpp: RenderSystem 등록 순서 조정
- SystemManager.h: 시스템 ID 목록에 RENDER_SYSTEM 추가
- WarpSystem.cpp: 워프 이벤트 처리 로직 추가 및 카메라 줌 이벤트 발생

+++ StarEffectSystem.cpp
@@ -26,7 +26,7 @@ namespace astro
                 // 반짝이는 효과
                 time += GetFrameTime() * twinkle;
                 size = maxSize * ((sinf(time) * 0.5f) + 0.5f);
-                bright = 127 + 127 * sinf(time);
+                bright = static_cast<int>(127 + 127 * sinf(time));
             }
         }
     }
